commit test

+++ PIDEV/src/main/java/pidev/elbey/Repositories/RolesRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pidev.elbey.Entities.Roles;
 
 @Repository
-public interface RolesRepo extends JpaRepository<Roles, Integer> {
+public interface RolesRepo extends JpaRepository<Roles, Long> {
     Roles findByName(String name);
 }
